Extract env var handling to new EnvSource class
